版本号更新为 v3.7.1 新增: 新增插件独立存储功能 "设置窗口-调试"中新增"打开配置文件存储目录"选项 优化断点续传操作
改动:
加快了任务数量很大时程序退出的速度

修复:
修复了正在运行下载任务时退出程序，任务状态保存错误的问题
修复了计算任务速度时可能会产生错误的问题
修复了下载AcPlay时程序出错的问题

+++ AcDown/AcDown/UI/Components/AcPlay.cs
@@ -36,7 +36,8 @@ namespace Kaedei.AcDown.UI.Components
 
          //删除所有旧版本文件
          string dir_acplay = Path.Combine(appdata, @"Kaedei\AcDown\UIComponents\AcPlay\");
-         Directory.Delete(dir_acplay, true);
+         if (Directory.Exists(dir_acplay))
+            Directory.Delete(dir_acplay, true);
 
 
          //建立文件夹
